moved diversity funtionality to garlinier and fixed glass node indexing
