+++ src/papers/GASkeleton.java
@@ -7,10 +7,9 @@ import general.HeuristicClasses.*;
 import general.SolutionClasses.SolutionConflictCounts;
 
 public class GASkeleton extends GCPHeuristic {
-    private GASkeletonSolution[] population;
-    private int k;
+    protected GASkeletonSolution[] population;
+    protected int k;
     private int iteration;
-    private double diversityThreshold = 20.0; // Paper uses low-diversity threshold around D < 20
 
     public GASkeleton(Options options) {
         super(options);
@@ -52,18 +51,16 @@ public class GASkeleton extends GCPHeuristic {
                 population[toReplace] = solution;
             }
 
-            // Track and check diversity every 100 iterations
-            if (iteration % 100 == 0) {
-                double diversity = calculatePopulationDiversity();
-                System.out.println("Iteration " + iteration + ": Diversity = " + String.format("%.2f", diversity));
-                if (diversity < diversityThreshold) {
-                    System.out.println("Population diversity too low (" + String.format("%.2f", diversity)
-                            + "), restarting population");
-                    InitPopulation();
-                    iteration = 0;
-                }
+            if (shouldRestartPopulation(iteration)) {
+                InitPopulation();
+                iteration = 0;
             }
         } while (this.report());
+    }
+
+    // Subclasses can hook this for heuristic-specific restart logic.
+    protected boolean shouldRestartPopulation(int iteration) {
+        return false;
     }
 
     public void InitPopulation() {
@@ -134,154 +131,6 @@ public class GASkeleton extends GCPHeuristic {
         }
 
         return count;
-    }
-
-    /**
-     * Calculates the partition distance between two colorings.
-     * This is the minimum number of elementary transformations (moving one node
-     * from one color class to another)
-     * needed to transform coloring1 into coloring2.
-     *
-     * Important: direct label mismatch is not a valid partition distance for graph
-     * coloring because color IDs are arbitrary. Two equivalent partitions may use
-     * different color labels (a permutation), which can make direct mismatch look
-     * large even when structures are identical. We therefore match color classes
-     * first (max overlap assignment) and then compute the remaining node moves.
-     */
-    public int calculatePartitionDistance(int[] coloring1, int[] coloring2) {
-        int numNodes = instance.getNumNodes();
-
-        // For very large k, avoid cubic assignment cost and fall back to direct mismatch.
-        if (k > 200) {
-            int mismatch = 0;
-            for (int i = 1; i <= numNodes; i++) {
-                if (coloring1[i] != coloring2[i]) {
-                    mismatch++;
-                }
-            }
-            return mismatch;
-        }
-
-        int[][] overlap = new int[k][k];
-        for (int i = 1; i <= numNodes; i++) {
-            int c1 = coloring1[i] - 1;
-            int c2 = coloring2[i] - 1;
-            // if c1 and c2 are valid colors 
-            if (c1 >= 0 && c1 < k && c2 >= 0 && c2 < k) {
-                overlap[c1][c2]++; // increment the overlap count for that pair of color classes
-            }
-        }
-
-        int maxOverlap = maxWeightAssignment(overlap);
-        return numNodes - maxOverlap;
-    }
-
-    // Hungarian algorithm (maximization form) over a square matrix.
-    private int maxWeightAssignment(int[][] weights) {
-        int n = weights.length;
-        if (n == 0) {
-            return 0;
-        }
-
-        // Convert maximization to minimization by subtracting every weight from maxW.
-        int maxW = 0;
-        for (int i = 0; i < n; i++) {
-            for (int j = 0; j < n; j++) {
-                if (weights[i][j] > maxW) {
-                    maxW = weights[i][j];
-                }
-            }
-        }
-
-        int[] rowPotential = new int[n + 1];
-        int[] colPotential = new int[n + 1];
-        int[] matchedRowForColumn = new int[n + 1];
-        int[] predecessorColumn = new int[n + 1];
-
-        for (int rowIndex = 1; rowIndex <= n; rowIndex++) {
-            matchedRowForColumn[0] = rowIndex; // mapping of col(index) to row(value)
-            int currCol = 0;
-            int[] colReduce = new int[n + 1];
-            boolean[] usedCols = new boolean[n + 1];
-            for (int colInd = 1; colInd <= n; colInd++) {
-                colReduce[colInd] = Integer.MAX_VALUE;
-                usedCols[colInd] = false;
-            }
-
-
-            do {
-                usedCols[currCol] = true;
-                int currRow = matchedRowForColumn[currCol];
-                int minDelta = Integer.MAX_VALUE; // min in colReduced
-                int nextColumn = 0;
-
-                // reduction for columns
-                for (int colInd = 1; colInd <= n; colInd++) {
-                    if (!usedCols[colInd]) {
-                        // to make maximization problem into minimization, we subtract weights from maxW
-                        int transformedCost = maxW - weights[currRow - 1][colInd - 1];
-                        int reducedCost = transformedCost - rowPotential[currRow] - colPotential[colInd];
-                        if (reducedCost < colReduce[colInd]) {
-                            colReduce[colInd] = reducedCost;
-                            predecessorColumn[colInd] = currCol;
-                        }
-                        if (colReduce[colInd] < minDelta) {
-                            minDelta = colReduce[colInd];
-                            nextColumn = colInd;
-                        }
-                    }
-                }
-
-                for (int colInd = 0; colInd <= n; colInd++) {
-                    if (usedCols[colInd]) {
-                        rowPotential[matchedRowForColumn[colInd]] += minDelta;
-                        colPotential[colInd] -= minDelta;
-                    } else {
-                        colReduce[colInd] -= minDelta;
-                    }
-                }
-                currCol = nextColumn;
-            } while (matchedRowForColumn[currCol] != 0);
-
-            do {
-                int previousColumn = predecessorColumn[currCol];
-                matchedRowForColumn[currCol] = matchedRowForColumn[previousColumn];
-                currCol = previousColumn;
-            } while (currCol != 0);
-        }
-
-        int total = 0;
-        for (int j = 1; j <= n; j++) {
-            int i = matchedRowForColumn[j];
-            if (i > 0) {
-                total += weights[i - 1][j - 1];
-            }
-        }
-        return total;
-    }
-
-    /**
-     * Calculates the average partition distance between all pairs of solutions in
-     * the population.
-     * This measures the diversity of the population.
-     */
-    public double calculatePopulationDiversity() {
-        int popSize = population.length;
-        if (popSize <= 1)
-            return 0.0;
-
-        double totalDistance = 0.0;
-        int pairs = 0;
-
-        for (int i = 0; i < popSize; i++) {
-            for (int j = i + 1; j < popSize; j++) {
-                int distance = calculatePartitionDistance(population[i].getColoring(), population[j].getColoring());
-                totalDistance += distance;
-                pairs++;
-            }
-        }
-
-        return pairs > 0 ? totalDistance / pairs : 0.0;
     }
 
     // returns the color partition with the most nodes
